Move invocation API related stuff from `reflector` to `java-commons` This is needed as this API is useful for various *magical* components of semi-standard API
