+++ java-commons/src/main/java/ru/progrm_jarvis/javacommons/invoke/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Common utilities for {@link java.lang.invoke Invoke API}
  */
-package ru.progrm_jarvis.reflector.invoke;
+package ru.progrm_jarvis.javacommons.invoke;
