+++ reflector/src/main/java/ru/progrm_jarvis/reflector/wrapper/invoke/InvokeDynamicFieldWrapper.java
@@ -8,7 +8,7 @@ import lombok.NonNull;
 import lombok.SneakyThrows;
 import lombok.experimental.FieldDefaults;
 import org.jetbrains.annotations.NotNull;
-import ru.progrm_jarvis.reflector.invoke.InvokeUtil;
+import ru.progrm_jarvis.javacommons.invoke.InvokeUtil;
 import ru.progrm_jarvis.reflector.wrapper.AbstractFieldWrapper;
 import ru.progrm_jarvis.reflector.wrapper.DynamicFieldWrapper;
